A slightly better arrangement
An explicit on-time initialization of the the HashBag class instance,
instead of repeated initializations...
