+++ lib/Hash.js
@@ -6,7 +6,6 @@
  */
 
 const crypto = require('crypto');
-const assert = require('assert');
 
 /**
  * Hash function to use
@@ -140,56 +139,11 @@ function hashBag(h1, h2) {
     return new HashValue(Buffer.from(result));
 }
 
-
-class HashBag {
-    constructor() {
-        this._values = {};
-    }
-
-    /**
-     * Init a new (hash) value to be accumulated for the blank node b.
-     * @param b: blank node identifier
-     * @param val: a hash value (instance of a HashValue class)
-     */
-    init(b, val) {
-        if(this._values[b] === undefined) {
-            this._values[b] = [val];
-        }
-    }
-
-    /**
-     * Add a new (hash) value to be accumulated for the blank node b
-     * @param b: blank node identifier
-     * @param val: a hash value (instance of a HashValue class)
-     */
-    addValue(b, val) {
-        assert.ok(this._values[b] !== undefined);
-        this._values[b].push(val)
-    }
-
-    /**
-     * Calculate the hash value related to the blank node 'b'. The value is calculated using the hashTuple function
-     * on the accumulated values, previous ordered.
-     *
-     * @param b: blank node identifier
-     * @return: corresponding hash value.
-     */
-    value(b) {
-        assert.ok(this._values[b] !== undefined);
-        return hashTuple(this._values[b].sort(HashValue.hashCompare))
-    }
-
-    get keys() {
-        return Object.keys(this._values);
-    }
-}
-
 /* =================================================================================== */
 
 module.exports = {
     HashValue,
     hashTerm,
     hashTuple,
-    hashBag,
-    HashBag
+    hashBag
 }
