+++ lib/HashTable.js
@@ -8,6 +8,7 @@
 const _      = require('underscore');
 const n3     = require('n3');
 const n3util = n3.Util;
+const assert = require('assert');
 
 const { HashValue, hashTerm, hashTuple, hashBag } = require('./Hash');
 
@@ -100,6 +101,13 @@ class HashTable {
             }
         }
         return undefined;
+    }
+
+    /**
+     * Get the bnode hashes
+     */
+    get bnode_hashes() {
+        return this._bnode_to_hash;
     }
 
     /**
@@ -213,7 +221,6 @@ class HashTable {
         return !_.some(this._hash_values, (tuple) => tuple.terms.length !== 1)
     }
 
-
     /**
      * Return list of blank node IDs ordered by their corresponding hash values.
      * This list can be used to assign canonical id-s to the blank nodes.
@@ -231,8 +238,60 @@ class HashTable {
     }
 }
 
+/*
+ * Structure to be used as a replacement of the hashBag function in the original paper:
+ * values for each blank node are collected in an array instead of making calculations,
+ * and these (hash) values are ordered increasingly and hashed as a tuple to provide
+ * a final hash value. This construction avoids the hash collision that may occur with
+ * the various implementations of the hashBag function that seems to lead to such
+ * collisions.
+ */
+class HashBag {
+    /**
+     * Take the blank node hash values from a HashTable to initialize the Internal
+     * arrays, one for each blank node.
+     *
+     * @param hashes: a HashTable instance
+     */
+    constructor(hashes) {
+        this._values = _.mapObject(hashes.bnode_hashes, (val,key) => [val])
+    }
+
+    /**
+     * Add a new (hash) value to be accumulated for the blank node b
+     * @param b: blank node identifier
+     * @param val: a hash value (instance of a HashValue class)
+     */
+    addValue(b, val) {
+        assert.ok(this._values[b] !== undefined);
+        this._values[b].push(val)
+    }
+
+    /**
+     * Calculate the hash value related to the blank node 'b'. The value is calculated using the hashTuple function
+     * on the accumulated values, previous ordered.
+     *
+     * @param b: blank node identifier
+     * @return: corresponding hash value.
+     */
+    value(b) {
+        assert.ok(this._values[b] !== undefined);
+        return hashTuple(this._values[b].sort(HashValue.hashCompare))
+    }
+
+    /**
+     * Get all the keys of the structure, ie, the blank node ids.
+     *
+     * @return: array of id-s
+     */
+    get keys() {
+        return _.keys(this._values);
+    }
+}
+
 /* =================================================================================== */
 
 module.exports = {
-    HashTable
+    HashTable,
+    HashBag
 }
