+++ lib/iso_canonical.js
@@ -9,9 +9,9 @@ const n3     = require('n3');
 const n3util = n3.Util;
 const assert = require('assert');
 
-const { HashValue, hashTerm, hashTuple, hashBag, HashBag }  = require('./Hash');
-const { HashTable }                                         = require('./HashTable');
-const { Dataset }                                           = require('./Dataset');
+const { HashValue, hashTerm, hashTuple, hashBag}  = require('./Hash');
+const { HashTable, HashBag }                      = require('./HashTable');
+const { Dataset }                                 = require('./Dataset');
 
 const debug = false;
 
@@ -45,13 +45,10 @@ function hashBnodes(G, hash) {
         let c = null;
         hashIm1  = hashI.clone();
         hashI    = hashIm1.clone();
-        if(ALTERNATIVE_BAG) hash_bag = new HashBag();
+        if(ALTERNATIVE_BAG) hash_bag = new HashBag(hashI);
         _.forEach(G.quads, (quad) => {
             if( n3util.isBlank(quad.subject) ) {
                 let b = quad.subject;
-                if(ALTERNATIVE_BAG) {
-                    hash_bag.init(b, hashI.getHash(b));
-                }
                 if( quad.graph ) {
                     c = hashTuple(hashIm1.getHash(quad.object), hashIm1.getHash(quad.predicate), hashIm1.getHash(quad.graph), '+')
                 } else {
@@ -65,9 +62,6 @@ function hashBnodes(G, hash) {
             }
             if( n3util.isBlank(quad.object) ) {
                 let b = quad.object;
-                if(ALTERNATIVE_BAG) {
-                    hash_bag.init(b, hashI.getHash(b));
-                }
                 if( quad.graph ) {
                     c = hashTuple(hashIm1.getHash(quad.subject), hashIm1.getHash(quad.predicate), hashIm1.getHash(quad.graph), '-')
                 } else {
@@ -81,9 +75,6 @@ function hashBnodes(G, hash) {
             }
             if( quad.graph && n3util.isBlank(quad.graph) ) {
                 let b = quad.graph;
-                if(ALTERNATIVE_BAG) {
-                    hash_bag.init(b, hashI.getHash(b));
-                }
                 c = hashTuple(hashIm1.getHash(quad.subject), hashIm1.getHash(quad.predicate), hashIm1.getHash(quad.object), '.')
                 if(ALTERNATIVE_BAG) {
                     hash_bag.addValue(b, c);
